# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@goslack.dev
    name: API Support
    url: https://github.com/heyrmi/goslack/issues
  description: A Slack-like collaboration platform API with real-time messaging, file
    sharing, and workspace management.
  license:
    name: MIT
    url: https://github.com/heyrmi/goslack/blob/main/LICENSE
  termsOfService: https://github.com/heyrmi/goslack
  title: GoSlack API
  version: "1.0"
paths:
...
  /files/{id}:
    delete:
      description: Delete a file (only file uploader can delete)
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: File deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid file ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only file uploader can delete
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete File
      tags:
      - files
    get:
      description: Retrieve file metadata by ID (requires appropriate access permissions)
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      - description: Workspace ID
        in: query
        name: workspace_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: File metadata
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid file ID or workspace ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied - workspace membership required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found or access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get File Metadata
      tags:
      - files
  /files/{id}/download:
    get:
      description: Download a file by ID (requires appropriate access permissions)
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "400":
          description: Invalid file ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found or access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download File
      tags:
      - files
  /files/message:
    post:
      consumes:
      - application/json
      description: Send a message with file attachment to channel or direct message
      parameters:
      - description: File message details
        in: body
        name: message
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: File message sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or file access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Workspace membership required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send File Message
      tags:
      - files
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file to a workspace, channel, or for direct messaging
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Workspace ID
        in: formData
        name: workspace_id
        required: true
        type: integer
      - description: Channel ID (for channel files)
        in: formData
        name: channel_id
        type: integer
      - description: Receiver User ID (for direct message files)
        in: formData
        name: receiver_id
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: File uploaded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request, file too large, or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied - workspace/channel membership required
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: File too large
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload File
      tags:
      - files
...
  /workspaces/{id}/files:
    get:
      description: List all files in a workspace (requires workspace membership)
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Files per page (default: 20, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of files with pagination
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid workspace ID or pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Workspace membership required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Workspace Files
      tags:
      - files
  /workspaces/{id}/files/stats:
    get:
      description: Get file statistics for a workspace (requires workspace membership)
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: File statistics
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid workspace ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Workspace membership required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get File Statistics
      tags:
      - files
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"